projects/agent-failure-taxonomy/projects/agent-failure-taxonomy/src/requirements.txt
# Agent Failure Reproduction Infrastructure Dependencies

# LLM Providers
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0

# Core utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Analysis and visualization (for later)
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
